@@ -1,6 +1,6 @@
 <?php
 
-namespace Laravelium\Sitemap;
+namespace Pelazio\Sitemap;
 
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Contracts\Support\DeferrableProvider;
@@ -9,16 +9,12 @@ use Illuminate\Contracts\Routing\ResponseFactory;
 
 class SitemapServiceProvider extends ServiceProvider implements DeferrableProvider
 {
-    /**
-     * Bootstrap the application events.
-     *
-     * @return void
-     */
-    public function boot()
-    {
-        $this->loadViewsFrom(__DIR__.'/../../views', 'sitemap');
 
-        $config_file = __DIR__.'/../../config/config.php';
+    public function boot(): void
+    {
+        $this->loadViewsFrom(__DIR__ . '/../../views', 'sitemap');
+
+        $config_file = __DIR__ . '/../../config/config.php';
 
         $this->mergeConfigFrom($config_file, 'sitemap');
 
@@ -27,20 +23,16 @@ class SitemapServiceProvider extends ServiceProvider implements DeferrableProvid
         ], 'config');
 
         $this->publishes([
-            __DIR__.'/../../views' => base_path('resources/views/vendor/sitemap'),
+            __DIR__ . '/../../views' => base_path('resources/views/vendor/sitemap'),
         ], 'views');
 
         $this->publishes([
-            __DIR__.'/../../public' => public_path('vendor/sitemap'),
+            __DIR__ . '/../../public' => public_path('vendor/sitemap'),
         ], 'public');
     }
 
-    /**
-     * Register the service provider.
-     *
-     * @return void
-     */
-    public function register()
+
+    public function register(): void
     {
         $this->app->bind('sitemap', function (Container $app) {
             $config = $app->make('config');
@@ -58,10 +50,7 @@ class SitemapServiceProvider extends ServiceProvider implements DeferrableProvid
         $this->app->alias('sitemap', Sitemap::class);
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function provides()
+    public function provides(): array
     {
         return ['sitemap', Sitemap::class];
     }
